Extracted the utils to a separate file
